Data/requirements.txt
pandas
sqlalchemy
psycopg2-binary
requests
python-dotenv
streamlit
plotly
openpyxl